Improve the putnbr

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: fgalaup <fgalaup@student.le-101.fr>        +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/10/07 16:16:00 by fgalaup      #+#   ##    ##    #+#       */
-/*   Updated: 2019/11/11 10:28:17 by fgalaup     ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/11/14 10:44:26 by fgalaup     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -58,7 +58,7 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 void	ft_putchar_fd(int fd, char c);
 void	ft_putstr_fd(int fd, char *s);
 void	ft_putendl_fd(int fd, char *s);
-void	ft_putnbr_fd(int fd, int n);
+void	ft_putnbr_fd(int fd, long long n);
 
 size_t	ft_count_digits(long long number);
 size_t	ft_count_udigits(unsigned long long number);
